* applied: http://www.ascentemu.com/forums/index.php?showtopic=10164&hl= * fixed where it was possible to break spirit of redemption
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@2641 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/game/ItemHandler.cpp
@@ -1802,7 +1802,7 @@ void WorldSession::HandleWrapItemOpcode(WorldPacket& recv_data)
 		return;
 	}
 
-	if(destitem_bagslot==0xFF && (destitem_slot >= EQUIPMENT_SLOT_START && destitem_slot <= EQUIPMENT_SLOT_END))
+	if(destitem_bagslot==(uint8)0xFF && (destitem_slot >= EQUIPMENT_SLOT_START && destitem_slot <= EQUIPMENT_SLOT_END))
 	{
 		_player->GetItemInterface()->BuildInventoryChangeError(src,dst,INV_ERR_EQUIPPED_CANT_BE_WRAPPED);
 		return;
